+++ custom_components/gwm_ora5/const.py
@@ -1,5 +1,5 @@
 """ORA 5 integration constants."""
 DOMAIN = "gwm_ora5"
 NAME = "GWM ORA 5"
-PLATFORMS = ["sensor", "binary_sensor", "button"]
+PLATFORMS = ["sensor", "binary_sensor", "button", "device_tracker", "lock"]
 POLL_SECONDS = 120